+++ hardhat/contracts/utils/testmapping.sol
@@ -4,18 +4,18 @@ pragma solidity ^0.8.11;
 contract testmapping {
     struct MyStruct {
         address subber;
-        uint something;
+        uint256 something;
     }
-    mapping(uint => MyStruct) data;
+    mapping(uint256 => MyStruct) data;
+
     constructor() {
         data[0] = MyStruct(address(1337), 7331);
         data[1] = MyStruct(address(4444), 5555);
     }
 
-    function d() public view returns(MyStruct[] memory rtn) {
+    function d() public view returns (MyStruct[] memory rtn) {
         rtn = new MyStruct[](2);
         rtn[0] = data[0];
         rtn[1] = data[1];
     }
-
 }
